Feat(nvim): add util file for global functions
- add functions for navigating location list

tags:
    location list, global functions, utils, refactoring
