+++ .config/nvim/lua/general/leader.lua
@@ -1,16 +1,4 @@
---https://sbulav.github.io/vim/neovim-telescope-github/
-function _G.save_and_exec()
-	vim.cmd([[silent! write]])
-	if vim.bo.filetype == "lua" then
-		require("plenary.reload").reload_module(vim.fn.expand("%"))
-		vim.cmd([[:luafile %]])
-	elseif vim.bo.filetype == "vim" then
-        vim.cmd([[:source %]])
-	else
-	end
-
-	print("Reloaded buffer : ",vim.fn.expand("%"))
-end
+require("utils.globalfunctions")
 
 vim.cmd([[
